requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Video download and processing
yt-dlp>=2023.12.30
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Audio transcription
openai-whisper>=20231117
torch>=2.1.0
torchaudio>=2.1.0

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0
requests>=2.31.0

# Web scraping and fact-checking
beautifulsoup4>=4.12.0
requests-html>=0.10.0
duckduckgo-search>=6.0.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
wordcloud>=1.9.2

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0
ipython>=8.18.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.2
markdown>=3.5.0

